Move tests to pixels
