+++ src/pydicom/pixels/encoders/__init__.py
@@ -1,5 +1,9 @@
+# Copyright 2008-2024 pydicom authors. See LICENSE file for details.
+
 from pydicom.pixels.encoders.base import (
     RLELosslessEncoder,
     JPEGLSLosslessEncoder,
     JPEGLSNearLosslessEncoder,
+    JPEG2000LosslessEncoder,
+    JPEG2000Encoder,
 )
