+++ tests/test_pillow_pixel_data.py
@@ -6,7 +6,8 @@ import pydicom
 from pydicom.data import get_testdata_file
 from pydicom.encaps import get_frame, generate_frames, encapsulate
 from pydicom.filereader import dcmread
-from pydicom.pixel_data_handlers.util import convert_color_space, get_j2k_parameters
+from pydicom.pixels.processing import convert_color_space
+from pydicom.pixels.utils import get_j2k_parameters
 from pydicom.uid import (
     JPEGBaseline8Bit,
     JPEGLosslessSV1,
